requirements.txt
diffusers
transformers
accelerate
scipy
safetensors
streamlit
streamlit-drawable-canvas
pyngrok
torch
numpy
Pillow